Conflicts:
	blockly_compressed.js
	blockly_uncompressed.js
	core/blockly.js

+++ core/blockly.js
@@ -43,7 +43,13 @@ goog.require('Blockly.FieldVariable');
 goog.require('Blockly.Generator');
 goog.require('Blockly.Msg');
 goog.require('Blockly.Procedures');
-goog.require('Blockly.Realtime');
+// Realtime is currently badly broken.  Stub it out.
+//goog.require('Blockly.Realtime');
+Blockly.Realtime = {
+  isEnabled: function() {return false;},
+  blockChanged: function() {},
+  doCommand: function(cmdThunk) {cmdThunk();}
+};
 goog.require('Blockly.Toolbox');
 goog.require('Blockly.TypeBlock');
 goog.require('Blockly.WidgetDiv');
@@ -56,6 +62,9 @@ goog.require('goog.color');
 goog.require('goog.events.KeyCodes');
 goog.require('goog.userAgent');
 
+
+// Turn off debugging when compiled.
+var CLOSURE_DEFINES = {'goog.DEBUG': false};
 
 /**
  * Required name space for SVG elements.
@@ -360,6 +369,7 @@ Blockly.onKeyDown_ = function(e) {
     // When focused on an HTML text input widget, don't trap any keys.
     return;
   }
+  var deleteBlock = false;
   if (e.keyCode == goog.events.KeyCodes.ESC) {
     // Pressing esc closes the context menu.
     Blockly.hideChaff();
@@ -375,8 +385,7 @@ Blockly.onKeyDown_ = function(e) {
     // Delete or backspace.
     try {
       if (Blockly.selected && Blockly.selected.isDeletable()) {
-        Blockly.hideChaff();
-        Blockly.selected.dispose(true, true);
+        deleteBlock = true;
       }
     } finally {
       // Stop the browser from going back to the previous page.
@@ -388,16 +397,16 @@ Blockly.onKeyDown_ = function(e) {
     var handled = false;
     if (Blockly.selected &&
         Blockly.selected.isDeletable() && Blockly.selected.isMovable()) {
-      Blockly.hideChaff();
       if (e.keyCode == goog.events.KeyCodes.C) {
         // 'c' for copy.
+        Blockly.hideChaff();
         Blockly.copy_(Blockly.selected);
         handled = true;
       } else if (e.keyCode ==goog.events.KeyCodes.X) {
         // 'x' for cut.
         Blockly.copy_(Blockly.selected);
-        Blockly.selected.dispose(true, true);
         handled = true;
+        deleteBlock = true;
       }
     }
     if (e.keyCode == goog.events.KeyCodes.V) {
@@ -445,6 +454,16 @@ Blockly.onKeyDown_ = function(e) {
 //      Blockly.selected.workspace.scrollToArea(Blockly.selected);
       Blockly.selectBlock(Blockly.selected);
       Blockly.selected.showContextMenu_(e);
+    }
+  }
+  if (deleteBlock) {
+    // Common code for delete and cut.
+    Blockly.hideChaff();
+    var heal = Blockly.dragMode_ != 2;
+    Blockly.selected.dispose(heal, true);
+    if (Blockly.highlightedConnection_) {
+      Blockly.highlightedConnection_.unhighlight();
+      Blockly.highlightedConnection_ = null;
     }
   }
 };
@@ -940,7 +959,9 @@ Blockly.longStop_ = function() {
  */
 Blockly.copy_ = function(block) {
   var xmlBlock = Blockly.Xml.blockToDom_(block);
-  Blockly.Xml.deleteNext(xmlBlock);
+  if (Blockly.dragMode_ != 2) {
+    Blockly.Xml.deleteNext(xmlBlock);
+  }
   // Encode start position in XML.
   var xy = block.getRelativeToSurfaceXY();
   xmlBlock.setAttribute('x', block.RTL ? -xy.x : xy.x);
@@ -1020,8 +1041,11 @@ Blockly.getMainWorkspaceMetrics_ = function() {
   if (this.toolbox_) {
     svgSize.width -= this.toolbox_.width;
   }
-  var viewWidth = svgSize.width - Blockly.Scrollbar.scrollbarThickness;
-  var viewHeight = svgSize.height - Blockly.Scrollbar.scrollbarThickness;
+  // Set the margin to match the flyout's margin so that the workspace does
+  // not jump as blocks are added.
+  var MARGIN = Blockly.Flyout.prototype.CORNER_RADIUS - 1;
+  var viewWidth = svgSize.width - MARGIN;
+  var viewHeight = svgSize.height - MARGIN;
   try {
     var blockBox = this.getCanvas().getBBox();
   } catch (e) {
